h4x: Add a new test suite for HTML parsing.
These tests are made to fail with the current implementation as a
demonstration of why a proper HTML parser is needed. These tests are a
bit contrived. I have limited experience with test-driven development.
These are just a proof of concept really.

I aim to also submit a patch that uses an HTML parser, but that will be
a separate branch because I'm having trouble figuring out the
appropriate one to use, and in the event that my solution isn't merged
there might still be a chance that the test gets merged.