h4x: Add Readonly to test dependencies.
It may not be worth the dependency just for the test when the tight
scope already makes it easy to see that the variables aren't being
changed... I probably am going overboard with Readonly anyway. I just
discovered it so I'm experimenting with its usage.

Just seeing if this fixes the travis-ci build.